+++ packages/apollo-link-http/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change log
 
 ### vNEXT
+- throw error on empty ExectionResult (missing)
 - support setting credentials, headers, and fetcherOptions in the setup of the link
 - removed sending of context to the server and allowed opt-in of sending extensions